+++ AllaganLib.GameSheets/Sheets/Rows/SatisfactionSupplyRow.cs
@@ -5,5 +5,5 @@ namespace AllaganLib.GameSheets.Sheets.Rows;
 
 public class SatisfactionSupplyRow : ExtendedSubrow<SatisfactionSupply, SatisfactionSupplyRow, SatisfactionSupplySheet>
 {
-    public SatisfactionNpcRow? Npc => this.Sheet.GetSatisfactionNpcSheet().GetRowBySupplyId(this.Base.RowId);
+    public SatisfactionNpcRow? Npc => this.Sheet.GetSatisfactionNpcSheet().GetRowBySupplyId(this.RowId);
 }